All attributes now configurable in XML.

+++ super-bar/src/main/java/com/miguelgaeta/super_bar/SuperBarConfig.java
@@ -62,7 +62,7 @@ class SuperBarConfig {
      *
      * @return Margin of the bar.
      */
-    public float getBarMargin() {
+    public int getBarMargin() {
 
         return this.barMargin;
     }
